# business-osint-gpt/openapi.yaml
openapi: 3.0.1
info:
  title: Business OSINT Assistant API
  description: Mock API for Business OSINT Assistant
  version: 1.0.0
servers:
  - url: http://localhost:5000
paths:
  /search_companies:
    get:
      summary: Search for companies using OpenCorporates (https://api.opencorporates.com/v0.4/companies/search)
      parameters:
        - in: query
          name: query
          schema:
            type: string
          required: true
          description: Company name or keyword
      responses:
        '200':
          description: A list of matching companies
          content:
            application/json:
              example:
                companies:
                  - name: Example Corp
                    jurisdiction: us_ca
                    company_number: '1234567'

  /get_market_trends:
    get:
      summary: Get Google Trends data via PyTrends (https://github.com/GeneralMills/pytrends)
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
          required: true
          description: Keyword to search
        - in: query
          name: geo
          schema:
            type: string
          required: false
          description: Geography code (e.g., US-NY)
      responses:
        '200':
          description: Time series of interest values
          content:
            application/json:
              example:
                trend:
                  - date: '2023-01-01'
                    interest: 50

  /get_local_competitors:
    get:
      summary: Get local competitors from Yelp Fusion API (https://api.yelp.com/v3/businesses/search)
      parameters:
        - in: query
          name: term
          schema:
            type: string
          required: true
          description: Business search term
        - in: query
          name: location
          schema:
            type: string
          required: true
          description: City or location
      responses:
        '200':
          description: A list of businesses
          content:
            application/json:
              example:
                businesses:
                  - name: Sample Bakery
                    rating: 4.5
                    address: 123 Main St

  /get_industry_data:
    get:
      summary: Get industry statistics from NAICS / Census API (https://api.census.gov/data/)
      parameters:
        - in: query
          name: naics
          schema:
            type: string
          required: true
          description: NAICS code
      responses:
        '200':
          description: Industry statistics data
          content:
            application/json:
              example:
                industry:
                  naics: '311811'
                  establishments: 1000
                  employment: 50000

  /get_customer_data:
    get:
      summary: Get demographic data from DataUSA (https://datausa.io/api/)
      parameters:
        - in: query
          name: location
          schema:
            type: string
          required: true
          description: Location or FIPS code
      responses:
        '200':
          description: Demographic statistics
          content:
            application/json:
              example:
                demographics:
                  population: 100000
                  median_age: 35.5

  /get_grants:
    get:
      summary: Search for grants on Grants.gov (https://www.grants.gov/grantsws/rest/opportunities/search)
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
          required: true
          description: Search keyword
      responses:
        '200':
          description: List of grant opportunities
          content:
            application/json:
              example:
                grants:
                  - title: Small Business Grant
                    agency: SBA
                    deadline: 2024-05-01

  /get_website_stack:
    get:
      summary: Get website technology stack via BuiltWith (https://builtwith.com/)
      parameters:
        - in: query
          name: url
          schema:
            type: string
          required: true
          description: Website URL
      responses:
        '200':
          description: Technology stack information
          content:
            application/json:
              example:
                technologies:
                  - name: WordPress
                    category: CMS

  /get_ad_transparency:
    get:
      summary: Search Meta Ad Library (https://www.facebook.com/ads/library/)
      parameters:
        - in: query
          name: query
          schema:
            type: string
          required: true
          description: Search keyword or page
      responses:
        '200':
          description: Ad transparency results
          content:
            application/json:
              example:
                ads:
                  - advertiser: Sample Brand
                    ad_count: 20